requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
Jinja2
